requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
whoosh==2.7.4
python-docx==1.1.0
PyPDF2==3.0.1
